+++ packages/brother_ptouch_label/src/brother_ptouch_label/printing.py
@@ -40,7 +40,7 @@ def print_text(
     margin_left: int | None = None,
     margin_right: int | None = None,
     fixed_width: int | None = None,
-    replicate: int = 1,
+    replicate: int | str = 1,
     threshold: int = 128,
     auto_cut: bool = True,
     half_cut: bool = False,